requirements.txt
flask
werkzeug
PyPDF2
sentence-transformers
transformers
torch
faiss-cpu
